+++ src/document/mod.rs
@@ -47,8 +47,9 @@ pub use command::{
     DocumentDropDatabaseRequest, DocumentDropIndexRequest, DocumentFindRequest,
     DocumentIndexRequest, DocumentInsertRequest, DocumentKillCursorRequest,
     DocumentListCollectionMetadataRequest, DocumentListCollectionsRequest,
-    DocumentListIndexesRequest, DocumentMutationScope, DocumentNamespace, DocumentReplaceRequest,
-    DocumentRequest, DocumentRequestId, DocumentUpdateRequest, MAX_DOCUMENT_INDEX_NAME_BYTES,
+    DocumentListDatabaseNamesRequest, DocumentListIndexesRequest, DocumentMutationScope,
+    DocumentNamespace, DocumentReplaceRequest, DocumentRequest, DocumentRequestId,
+    DocumentUpdateRequest, MAX_DOCUMENT_INDEX_NAME_BYTES,
 };
 pub use distinct::DocumentDistinct;
 pub use error::{BsonError, BsonErrorContext, BsonErrorKind, BsonResult, DocumentCursorError};